Add the ability for an order to specify the quantity of a target token desired
Orders now have a sourceQuantity and targetQuantity field, one of which must be null. This allows orders to specify the amount of a sourceToken to use to fund the order or the amount of a targetToken they want after the order has executed. Future commits will check the user's balances to ensure they have enough sourceToken to buy the specified amount of targetToken.